refactor: JPA Entity와 repository 패키지 변경
