+++ out-adapter-persistence-jpa/src/main/java/gohigher/application/entity/ApplicationRepository.java
@@ -1,4 +1,4 @@
-package gohigher.application;
+package gohigher.application.entity;
 
 import org.springframework.data.jpa.repository.JpaRepository;
