+++ out-adapter-persistence-jpa/src/main/java/gohigher/recruitment/entity/RecruitmentProcessJpaEntity.java
@@ -1,4 +1,4 @@
-package gohigher.recruitment;
+package gohigher.recruitment.entity;
 
 import java.time.LocalDateTime;
